Made pathfinding threadsafe
- Added option to disable optimisation, and disabled it by default.
Reason: There's functionality that instead of news and deletes fetches
empty PathNodes from a table. This uses more memory but doubles the
pathfinding speed. Problem is that I need to add critical sections on
these probably. Also, we will have to increase the table size, it allows
for 2<<15 objects, but if there are >200 minions, it will probably try
to fetch outside of the table. if they all make paths.
- Moved Pathfinding variables into Path class which is now the new
return of the getPath.
- Added error to pathreturn. (Reason why path wasn't created or zero)